# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: A-Level Revision Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user-subjects:
    get:
      operationId: listCurrentUserSubjects
      tags: [userSubjects]
      summary: List the authenticated user's selected subjects
      description: Returns only the caller's durable memberships with shared metadata-only subject references.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user's selected subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSubjectMembership"
        "401":
          description: Missing or invalid Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
    put:
      operationId: replaceCurrentUserSubjects
      tags: [userSubjects]
      summary: Atomically replace the authenticated user's selected subjects
      description: |
        Replaces the caller's membership set with 1–5 distinct catalogue subjects.
        Existing tasks and other owned activity are not deleted.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSubjectSelectionInput"
      responses:
        "200":
          description: Replaced current user's selected subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSubjectMembership"
        "400":
          description: Invalid subject selection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "401":
          description: Missing or invalid Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /user-subjects/{subjectId}/assessment-route:
    put:
      operationId: assignCurrentUserSubjectAssessmentRoute
      tags: [userSubjects]
      summary: Assign or change the assessment route for one membership
      description: |
        Intentional authenticated mutation. Stays within the existing pinned
        syllabus_version_id. Does not run on page view. Legacy null-route
        remediation and Settings route changes use this endpoint.
      security:
        - bearerAuth: []
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignAssessmentRouteInput"
      responses:
        "200":
          description: Updated membership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSubjectMembership"
        "400":
          description: Invalid route or option selection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "401":
          description: Missing or invalid Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "404":
          description: Membership not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth access token (Authorization Bearer) for authenticated caller-owned routes.
  schemas:
    ErrorMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ExamSittingSeries:
      type: string
      enum:
        - Feb/Mar
        - May/June
        - Oct/Nov
    IntendedExamSession:
      type: object
      description: >
        Structured intended final exam sitting. Required at runtime when the
        request creates a new membership. Optional for retained-only or
        removal-only Settings replacement. Clients never send a version id.
      properties:
        year:
          type: integer
          minimum: 1000
          maximum: 9999
        series:
          $ref: "#/components/schemas/ExamSittingSeries"
      required:
        - year
        - series
      additionalProperties: false
...
    SubjectSessionOverride:
      type: object
      properties:
        subjectId:
          type: integer
          minimum: 1
        year:
          type: integer
          minimum: 1000
          maximum: 9999
        series:
          $ref: "#/components/schemas/ExamSittingSeries"
      required:
        - subjectId
        - year
        - series
      additionalProperties: false
    SubjectRouteAssignmentInput:
      type: object
      description: |
        Version-scoped route + study-option selection for one subject.
        Server revalidates route/options against the resolved published route set.
      properties:
        subjectId:
          type: integer
          minimum: 1
        routeId:
          type: integer
          minimum: 1
        optionIds:
          type: array
          items:
            type: integer
            minimum: 1
          uniqueItems: true
      required:
        - subjectId
        - routeId
        - optionIds
      additionalProperties: false
...
    UserSubjectSelectionInput:
      type: object
      properties:
        subjectIds:
          type: array
          minItems: 1
          maxItems: 5
          uniqueItems: true
          items:
            type: integer
            minimum: 1
        intendedExamSession:
          $ref: "#/components/schemas/IntendedExamSession"
        subjectSessionOverrides:
          type: array
          items:
            $ref: "#/components/schemas/SubjectSessionOverride"
        routeAssignments:
          type: array
          items:
            $ref: "#/components/schemas/SubjectRouteAssignmentInput"
      required:
        - subjectIds
      additionalProperties: false
    UserSubjectMembership:
      type: object
      properties:
        subject:
          $ref: "#/components/schemas/SubjectReference"
        syllabusVersion:
          $ref: "#/components/schemas/UserSubjectSyllabusVersion"
        assessmentRouteId:
          type: ["integer", "null"]
          description: |
            Canonical version-scoped assessment route. Null for legacy memberships
            until the member intentionally assigns a route.
        intendedExamSession:
          allOf:
            - $ref: "#/components/schemas/IntendedExamSession"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - subject
        - syllabusVersion
        - assessmentRouteId
        - intendedExamSession
        - createdAt
        - updatedAt
    SubjectReference:
      type: object
      description: Shared metadata for a subject selected by the authenticated caller.
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        color:
          type: string
        topicsTotal:
          type: integer
          description: Count of shared syllabus topics for this subject
      required:
        - id
        - name
        - code
        - color
        - topicsTotal
    UserSubjectSyllabusVersion:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
        examBoard:
          type: string
        qualification:
          type: string
      required:
        - id
        - label
        - examBoard
        - qualification
...
    AssignAssessmentRouteInput:
      type: object
      properties:
        routeId:
          type: integer
          minimum: 1
        optionIds:
          type: array
          items:
            type: integer
            minimum: 1
          uniqueItems: true
      required:
        - routeId
        - optionIds
      additionalProperties: false
...