# lib/api-zod/src/generated/types/assignAssessmentRouteInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * A-Level Revision Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AssignAssessmentRouteInput {
  /** @minimum 1 */
  routeId: number;
  /** @items.minimum 1 */
  optionIds: number[];
}