# lib/api-zod/src/generated/types/userSubjectMembership.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * A-Level Revision Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { IntendedExamSession } from './intendedExamSession';
import type { SubjectReference } from './subjectReference';
import type { UserSubjectSyllabusVersion } from './userSubjectSyllabusVersion';

export interface UserSubjectMembership {
  subject: SubjectReference;
  syllabusVersion: UserSubjectSyllabusVersion;
  /**
     * Canonical version-scoped assessment route. Null for legacy memberships
     * until the member intentionally assigns a route.
     * @nullable
     */
  assessmentRouteId: number | null;
  intendedExamSession: IntendedExamSession | null;
  createdAt: string;
  updatedAt: string;
}